use sublime project function
Projects is an integral part of my workflow in Sublime Text. A project
is just a Sublime workspace in which your folders are open and stored in
the sidebar. This helps since you can define a project and add folders
to it, and be able to switch between folders quickly.Using projects, you
will no longer have to go digging in Windows Explorer or Finder to get
the project you want and drag it into Sublime.